requirements.txt
requests
pillow
pyperclip
rawpy
imageio
